+++ src/main/java/edu/kpi/testcourse/bigtable/BigTable.java
@@ -4,6 +4,8 @@ package edu.kpi.testcourse.bigtable;
 // ⚠️ base, implementation of key-value storage, etc)
 
 import com.google.gson.JsonObject;
+import io.micronaut.security.token.jwt.render.BearerAccessRefreshToken;
+import java.util.Set;
 
 /**
  * BigTable is a key-value storage...
@@ -12,6 +14,10 @@ public interface BigTable {
   void putUser(String key, JsonObject value);
 
   void putUrl(String key, String value);
+
+  void putToken(String token);
+
+  Set<String> getTokens();
 
   Integer getUrlId();
