+++ src/main/java/edu/kpi/testcourse/rest/UsersController.java
@@ -1,6 +1,7 @@
 package edu.kpi.testcourse.rest;
 
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import com.nimbusds.jose.shaded.json.JSONObject;
 import edu.kpi.testcourse.Main;
 import edu.kpi.testcourse.bigtable.BigTableImpl;
@@ -11,6 +12,7 @@ import io.micronaut.http.HttpResponse;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.annotation.Body;
 import io.micronaut.http.annotation.Controller;
+import io.micronaut.http.annotation.Header;
 import io.micronaut.http.annotation.Post;
 import io.micronaut.http.client.RxHttpClient;
 import io.micronaut.http.client.annotation.Client;
@@ -19,17 +21,18 @@ import io.micronaut.security.authentication.UsernamePasswordCredentials;
 import io.micronaut.security.rules.SecurityRule;
 import io.micronaut.security.token.jwt.render.BearerAccessRefreshToken;
 import java.util.Map;
+import java.util.UUID;
 import javax.inject.Inject;
 
 /**
  * Controller for login.
  */
-@Secured(SecurityRule.IS_ANONYMOUS)
 @Controller("/users")
-public class UserSignInController {
+public class UsersController {
   @Inject
   @Client("/")
   RxHttpClient client;
+
   /**
    * in field "body" in Postman you should provide user email and password like
    * {
@@ -37,7 +40,7 @@ public class UserSignInController {
    *   "password" : "yourPassword"
    * }.
    */
-
+  @Secured(SecurityRule.IS_ANONYMOUS)
   @Post(value = "/signin",
       consumes = MediaType.APPLICATION_JSON,
       produces = MediaType.APPLICATION_JSON)
@@ -54,21 +57,58 @@ public class UserSignInController {
     UsernamePasswordCredentials credentials =
         new UsernamePasswordCredentials(user.getEmail(), user.getPassword());
 
-    HttpResponse<BearerAccessRefreshToken> httpResponse;
+    HttpResponse<String> httpResponse;
+
     boolean response = UserActions.findUserByEmail(user.getEmail());
     if (response) {
       boolean isPasswordValid = UserActions.checkPassword(user.getEmail(), user.getPassword());
       if (isPasswordValid) {
         HttpRequest request = HttpRequest.POST("/login", credentials);
-        httpResponse = client.toBlocking().exchange(request, BearerAccessRefreshToken.class);
+        httpResponse = client.toBlocking().exchange(request, String.class);
       } else {
         return HttpResponse.badRequest("Invalid password");
       }
     } else {
       return HttpResponse.notFound("User with email \"" + user.getEmail() + "\" was not found.");
     }
-
+    JsonObject jsonObject = JsonParser.parseString(httpResponse.body()).getAsJsonObject();
+    BigTableImpl.tokens.add(jsonObject.get("access_token").toString().replace("\"", ""));
     return httpResponse;
+  }
+
+  @Secured(SecurityRule.IS_AUTHENTICATED)
+  @Post(value = "/signout", consumes = MediaType.APPLICATION_JSON)
+  public HttpResponse signOut(@Header("Authorization") String token) {
+    if(BigTableImpl.tokens.contains(token.split(" ")[1])) {
+      token = token.split(" ")[1]; // delete part 'Bearer '
+      BigTableImpl.tokens.remove(token);
+      return HttpResponse.noContent();
+    } else {
+      return HttpResponse.unauthorized();
+    }
+  }
+
+  /**
+   * in field "body" in Postman you should provide user email and password like
+   * {
+   *   "email" : "yourEmail",
+   *   "password" : "yourPassword"
+   * }
+   * then you will see response that your user is created.
+   */
+  @Secured(SecurityRule.IS_ANONYMOUS)
+  @Post(value = "/signup",
+    consumes = MediaType.APPLICATION_JSON,
+    produces = MediaType.APPLICATION_JSON)
+  public HttpResponse signUp(@Body JSONObject object) {
+    User user = Main.getGson().fromJson(object.toJSONString(), User.class);
+    user.setUuid(UUID.randomUUID().toString());
+    boolean response = UserActions.createUser(user);
+    if (response) {
+      return HttpResponse.created("User with email \"" + user.getEmail() + "\" was created.");
+    } else {
+      return HttpResponse.badRequest("User with email \"" + user.getEmail() + "\" already exists.");
+    }
   }
 }
